# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Hotel {
  id: number;
  name: string;
  location: string;
  city: string;
  address: string;
  description: string;
  rating: number;
  imageUrl: string;
  amenities: string[];
  priceFrom: number;
  totalRooms: number;
  availableRooms: number;
  phone: string;
  email: string;
}

export interface HotelSummary {
  hotelId: number;
  totalRooms: number;
  availableRooms: number;
  occupancyRate: number;
  totalBookings: number;
  popularRoomType: string;
  avgRating: number;
}

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  available: "available",
  reserved: "reserved",
  occupied: "occupied",
  cleaning: "cleaning",
  maintenance: "maintenance",
} as const;

export interface Room {
  id: number;
  hotelId: number;
  roomNumber: string;
  type: string;
  description: string;
  pricePerNight: number;
  capacity: number;
  imageUrl: string;
  amenities: string[];
  isAvailable: boolean;
  status: RoomStatus;
  floor: number;
  view: string;
}

export interface BookingRequest {
  roomId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  specialRequests?: string;
}

export interface Booking {
  id: number;
  roomId: number;
  hotelId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  specialRequests?: string;
  status: string;
  source: string;
  externalRef?: string | null;
  checkedInAt?: string | null;
  checkedOutAt?: string | null;
  guestId?: number | null;
  totalPrice: number;
  createdAt: string;
  room?: Room;
  hotel?: Hotel;
}

export type RoomStatusUpdateStatus =
  (typeof RoomStatusUpdateStatus)[keyof typeof RoomStatusUpdateStatus];

export const RoomStatusUpdateStatus = {
  available: "available",
  reserved: "reserved",
  occupied: "occupied",
  cleaning: "cleaning",
  maintenance: "maintenance",
} as const;

export interface RoomStatusUpdate {
  status: RoomStatusUpdateStatus;
}

export interface Guest {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  notes?: string | null;
  totalBookings: number;
  sources: string[];
  lastStayAt?: string | null;
  createdAt: string;
}

export type GuestWithHistory = Guest & {
  bookings: Booking[];
};

export interface OtaIngestRequest {
  /** OTA's unique booking reference (used for idempotency) */
  externalRef: string;
  roomId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  totalPrice: number;
  specialRequests?: string;
}