# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxury Hotel System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/check-in:
    post:
      operationId: checkInBooking
      tags: [bookings]
      summary: Mark a booking as checked-in (guest arrived)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking checked-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/check-out:
    post:
      operationId: checkOutBooking
      tags: [bookings]
      summary: Mark a booking as checked-out (guest departed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking checked-out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest profile with full booking history (merged across web + OTA)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest with bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestWithHistory"
...
components:
  schemas:
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        city:
          type: string
        address:
          type: string
        description:
          type: string
        rating:
          type: number
        imageUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
        priceFrom:
          type: number
        totalRooms:
          type: integer
        availableRooms:
          type: integer
        phone:
          type: string
        email:
          type: string
      required:
        - id
        - name
        - location
        - city
        - address
        - description
        - rating
        - imageUrl
        - amenities
        - priceFrom
        - totalRooms
        - availableRooms
        - phone
        - email
...
    Room:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        roomNumber:
          type: string
        type:
          type: string
        description:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        imageUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        status:
          type: string
          enum: [available, reserved, occupied, cleaning, maintenance]
        floor:
          type: integer
        view:
          type: string
      required:
        - id
        - hotelId
        - roomNumber
        - type
        - description
        - pricePerNight
        - capacity
        - imageUrl
        - amenities
        - isAvailable
        - status
        - floor
        - view
    BookingRequest:
      type: object
      properties:
        roomId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        specialRequests:
          type: string
      required:
        - roomId
        - guestName
        - guestEmail
        - guestPhone
        - checkInDate
        - checkOutDate
        - numberOfGuests
    Booking:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        hotelId:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkInDate:
          type: string
        checkOutDate:
          type: string
        numberOfGuests:
          type: integer
        specialRequests:
          type: string
        status:
          type: string
        source:
          type: string
        externalRef:
          type: string
          nullable: true
        checkedInAt:
          type: string
          nullable: true
        checkedOutAt:
          type: string
          nullable: true
        guestId:
          type: integer
          nullable: true
        totalPrice:
          type: number
        createdAt:
          type: string
        room:
          $ref: "#/components/schemas/Room"
        hotel:
          $ref: "#/components/schemas/Hotel"
      required:
        - id
        - roomId
        - hotelId
        - guestName
        - guestEmail
        - guestPhone
        - checkInDate
        - checkOutDate
        - numberOfGuests
        - status
        - source
        - totalPrice
        - createdAt
...
    Guest:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        notes:
          type: string
          nullable: true
        totalBookings:
          type: integer
        sources:
          type: array
          items:
            type: string
        lastStayAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - fullName
        - email
        - phone
        - totalBookings
        - sources
        - createdAt
    GuestWithHistory:
      allOf:
        - $ref: "#/components/schemas/Guest"
        - type: object
          properties:
            bookings:
              type: array
              items:
                $ref: "#/components/schemas/Booking"
          required:
            - bookings
...