# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */
import type { Hotel } from "./hotel";
import type { Room } from "./room";

export interface Booking {
  id: number;
  roomId: number;
  hotelId: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkInDate: string;
  checkOutDate: string;
  numberOfGuests: number;
  specialRequests?: string;
  status: string;
  source: string;
  externalRef?: string | null;
  checkedInAt?: string | null;
  checkedOutAt?: string | null;
  guestId?: number | null;
  totalPrice: number;
  createdAt: string;
  room?: Room;
  hotel?: Hotel;
}