# lib/api-zod/src/generated/types/guest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */

export interface Guest {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  notes?: string | null;
  totalBookings: number;
  sources: string[];
  lastStayAt?: string | null;
  createdAt: string;
}