# lib/api-zod/src/generated/types/guestWithHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";
import type { Guest } from "./guest";

export type GuestWithHistory = Guest & {
  bookings: Booking[];
};