# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Hotel System API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomStatus } from "./roomStatus";

export interface Room {
  id: number;
  hotelId: number;
  roomNumber: string;
  type: string;
  description: string;
  pricePerNight: number;
  capacity: number;
  imageUrl: string;
  amenities: string[];
  isAvailable: boolean;
  status: RoomStatus;
  floor: number;
  view: string;
}